backend/requirements.txt
absl-py==2.0.0
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
bson==0.5.10
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
decorator==4.4.2
Deprecated==1.2.14
dnspython==2.4.2
ffmpeg-python==0.2.0
filelock==3.13.1
firebase-admin==6.2.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
flatbuffers==23.5.26
fonttools==4.44.0
fsspec==2023.10.0
future==0.18.3
gcloud==0.18.3
gensim==4.3.2
google-api-core==2.14.0
google-api-python-client==2.107.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.0
grpcio==1.59.2
grpcio-status==1.59.2
httplib2==0.22.0
huggingface-hub==0.19.4
idna==3.4
imageio==2.32.0
imageio-ffmpeg==0.4.9
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jwcrypto==1.5.0
kiwisolver==1.4.5
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.1
mediapipe==0.10.8
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
oauth2client==4.1.3
opencv-contrib-python==4.8.1.78
packaging==23.2
pandas==2.1.3
Pillow==10.0.1
playwright==1.39.0
proglog==0.1.10
proto-plus==1.22.3
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pyee==11.0.1
PyJWT==2.8.0
pymongo==4.6.0
pyparsing==3.1.1
Pyrebase4==4.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jwt==4.1.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
smart-open==6.4.0
sounddevice==0.4.6
soupsieve==2.5
SpeechRecognition==3.10.0
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.35.2
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
uuid==1.30
Werkzeug==3.0.1
wrapt==1.16.0
